move validateChecksum methods to WasapiResponseValidator class
